Revert "Merge branch 'master' of https://github.com/entdark/jaMME"
This reverts commit 69606c306745361c74e8ebbe77ea901f33e045ed, reversing
changes made to 3f1fdd1de59f1e4317c20aec7f850bb11b56e4f0.

+++ codemp/client/snd_mix.h
@@ -19,7 +19,7 @@ typedef struct mixSound_s {
 
 typedef struct {
 	sfxHandle_t		handle;
-	int64_t			index;
+	int				index;
 	vec3_t			origin;
 	short			entNum;
 	char			entChan;
